moved scrapy/selector -> scrapyselectors, added gitignore
