--- scrapy/selector/__init__.py
@@ -1,5 +1,0 @@
-"""
-Selectors
-"""
-from scrapy.selector.unified import *
-from scrapy.selector.lxmlsel import *
