+++ scrapyselectors/__init__.py
@@ -1,0 +1,5 @@
+"""
+Selectors
+"""
+from scrapyselectors.unified import *
+from scrapyselectors.lxmlsel import *
