+++ scrapyselectors/lxmldocument.py
