+++ scrapyselectors/lxmlsel.py
