+++ tests/test_selector_lxmldocument.py
@@ -1,5 +1,5 @@
 import unittest
-from scrapy.selector.lxmldocument import LxmlDocument
+from scrapyselectors.lxmldocument import LxmlDocument
 from scrapy.http import TextResponse, HtmlResponse
